third_party/WebKit/LayoutTests/virtual/enable_asmjs/http/tests/asmjs/asm-warnings-expected.txt

CONSOLE WARNING: line 13: Invalid asm.js: Invalid type annotation - forbidden literal.
CONSOLE WARNING: line 5: Invalid asm.js: Invalid initializer for foreign import - unrecognized annotation.
